refactor : <ktp_task_controller> modularize task processing logics
