+++ ktp_task_controller/launch/ktp_task_controller.launch.py
@@ -1,3 +1,4 @@
+import os;
 from launch import LaunchDescription;
 from launch_ros.actions import Node;
 from ament_index_python.packages import get_package_share_directory;
@@ -8,13 +9,14 @@ def generate_launch_description() -> LaunchDescription:
 
     package_name: str = "ktp_task_controller";
     package_shared_directory: str = get_package_share_directory(package_name);
+    parameters: list[str] = [os.path.join(package_shared_directory, "config", "common_config.yaml")]
 
     converter_node: Node = Node(
         package=package_name,
         executable=package_name,
         name=package_name,
         output="screen",
-        parameters=[]
+        parameters=parameters
     );
 
     ld.add_action(converter_node)
